Add command for adding transactions

+++ src/main/java/com/github/adrjo/transactions/Transaction.java
@@ -1,4 +1,4 @@
 package com.github.adrjo.transactions;
 
-public record Transaction(String desc, float amt, long timestamp) {
+public record Transaction(String desc, double amt, long timestamp) {
 }
